+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+- Added Get-PnPTenantId to retrieve the current tenant id.
+
+### Changed
+
+### Contributors
+
 ## [3.15.1911.0]
 
 ### Added